DRTVWR-418: Force $LL_BUILD_RELEASE switches into Mac build.
Setting CFLAGS and CXXFLAGS was simply not working: the Configure script
totally ignored those values. Its output reports CFLAG (not CFLAGS), but
setting CFLAG didn't work any better. This caused openssl to be built for OS X
10.11, not 10.9 as specified in LL_BUILD_RELEASE.

What works is to pass in the relevant switches on the Configure command line.
Any unrecognized argument starting with a dash, it passes through to the
compiler.

Unfortunately this fails for switches specified as separate arguments. Our Mac
LL_BUILD_RELEASE variable contains (e.g.) -iwithsysroot /path/to/sysroot. This
caused various parts of the build to error out when they didn't know how to
handle /path/to/sysroot.

Instead of backtracking to impose a single-argument requirement on EVERY
switch in build-variables and viewer-build-variables, introduce a pass over
LL_BUILD_RELEASE switches to produce -iwithsysroot=/path/to/sysroot. That
appears to cause openssl's Configure script to do what we want.